Task 0

+++ 0x06-python-classes/0-square.py
@@ -5,5 +5,5 @@ class Square:
     Represents a square
 
     Attributes:
-        length (float): length of sides
+        length (int): length of sides
     """
